Split the Ollama API implementation to a separated file
